requirements.txt
# ============================================================
# Career Navigator v2 — Python Dependencies
# Install: pip install -r requirements.txt
# ============================================================

# ── Web Framework ────────────────────────────────────────────
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.32.0         # uvloop + httptools for performance

# ── Async Database ───────────────────────────────────────────
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0                   # async PostgreSQL driver
pgvector>=0.3.6                   # pgvector Python bindings for SQLAlchemy
alembic>=1.14.0                   # database migrations

# ── Auth ─────────────────────────────────────────────────────
python-jose[cryptography]>=3.3.0  # JWT tokens
bcrypt>=4.2.0                     # direct bcrypt — bypasses passlib 1.7/4.x bug
python-multipart>=0.0.12          # file uploads

# ── Settings ─────────────────────────────────────────────────
pydantic-settings>=2.6.0
pydantic[email]>=2.10.0

# ── Async HTTP (job APIs) ────────────────────────────────────
httpx>=0.28.0

# ── Task Queue (Heavy Worker) ────────────────────────────────
celery[redis]>=5.4.0
redis>=5.2.0
flower>=2.0.0                     # Celery monitoring UI at :5555

# ── Background Scheduling ────────────────────────────────────
APScheduler>=3.10.4

# ── Resume Parsing ───────────────────────────────────────────
pdfplumber>=0.11.0

# ── Semantic Skill Matching (384-dim vectors) ────────────────
# ⚠️ CAPACITY FLAG: Loads ~600MB RAM per worker process.
# With --workers 4: ~2.4GB RAM required. Reduce to --workers 1 if needed.
sentence-transformers>=3.3.0

# ── Agentic Framework (Phase 1 Inference Agent) ─────────────
langchain>=0.3.0
langchain-anthropic>=0.3.0
langchain-openai>=0.3.0
anthropic>=0.40.0                 # Direct SDK for Phase 3 Artifact Sovereign
openai>=1.58.0

# ── PDF Generation (Phase 3 Study Vault) ────────────────────
# ⚠️ CAPACITY FLAG: CPU-intensive. For > 10 concurrent PDF renders:
# Celery queue is already architected. For > 200/day: AWS Lambda + S3.
reportlab>=4.2.0
Pillow>=11.0.0

# ── Testing ──────────────────────────────────────────────────
pytest>=8.3.0
pytest-asyncio>=0.24.0
